render camera to texture

+++ scripts/rust_setup.sh
@@ -1,15 +1,13 @@
+
 # 设置 rustup 国内源
-export RUSTUP_DIST_SERVER="https://mirrors.ustc.edu.cn/rust-static"
-export RUSTUP_UPDATE_ROOT="https://mirrors.ustc.edu.cn/rust-static/rustup"
+rust_env_setup() {
+    export RUSTUP_DIST_SERVER="https://mirrors.ustc.edu.cn/rust-static"
+    export RUSTUP_UPDATE_ROOT="https://mirrors.ustc.edu.cn/rust-static/rustup"
 
-export RUSTUP_HOME=$HOME/.rustup
-export CARGO_HOME=$HOME/.cargo
-export PATH=$PATH:$CARGO_HOME/bin
-
-if [[ -f $CARGO_HOME/env ]]; then
-    source $CARGO_HOME/env
-    exit 0
-fi
+    export RUSTUP_HOME=$HOME/.rustup
+    export CARGO_HOME=$HOME/.cargo
+    export PATH=$PATH:$CARGO_HOME/bin
+}
 
 detect_osname() {
     local os_name="unknown"
@@ -44,41 +42,53 @@ detect_osname() {
         fi
     fi
 }
-# 1. 卸载 brew 安装的 rust
-os=$(detect_osname)
 
-if [[ $os == "macOS" ]]; then
-    brew uninstall rust
-elif [[ $os == "Linux" ]]; then
-    sudo apt remove rust
-else
-    choco uninstall rust
-fi
+rust_install() {
+    if [[ -f $CARGO_HOME/env ]]; then
+        source $CARGO_HOME/env
+        return 0
+    fi
 
-# 2. 安装 rustup（如果还没装）
-if command -v rustup >/dev/null 2>&1; then
-    echo "rustup is installed at: $(which rustup)"
-else
-    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
-fi
+    # 1. 卸载 brew 安装的 rust
+    os=$(detect_osname)
 
-# 3. 激活环境
-if [[ ! -d $HOME/.cargo ]]; then
-    mkdir -p $HOME/.cargo
-fi
+    if [[ $os == "macOS" ]]; then
+        brew uninstall rust
+    elif [[ $os == "Linux" ]]; then
+        sudo apt remove rust
+    else
+        choco uninstall rust
+    fi
 
-# 4. 设置默认工具链
-rustup default stable
+    # 2. 安装 rustup（如果还没装）
+    if command -v rustup >/dev/null 2>&1; then
+        echo "rustup is installed at: $(which rustup)"
+    else
+        curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+    fi
 
-# 5. 验证
-if [[ ! -f "$HOME/.cargo/env" ]]; then
-    rustup-init
-fi
-rustc --version
-cargo --version
+    # 3. 激活环境
+    if [[ ! -d $HOME/.cargo ]]; then
+        mkdir -p $HOME/.cargo
+    fi
 
-rustup target add aarch64-linux-android
+    # 4. 设置默认工具链
+    rustup default stable
 
-if [[ -f "$HOME/.cargo/env" ]]; then
-    source "$HOME/.cargo/env"
-fi
+    # 5. 验证
+    if [[ ! -f "$HOME/.cargo/env" ]]; then
+        rustup-init
+    fi
+    rustc --version
+    cargo --version
+
+    rustup target add aarch64-linux-android
+
+    if [[ -f "$HOME/.cargo/env" ]]; then
+        source "$HOME/.cargo/env"
+    fi
+
+}
+
+rust_env_setup
+rust_install
